Augment users data; add useful information

This returns the user dataframe, with added columns containing distance between that user and a given facility - IDs are generated for IDs and facilities that correspond to their row number.

augment_user(facilities_selected, existing_facilities, existing_users)

Arguments

facilities_selected
dataframe of facilities selected, obtained from extract_facility_selected

existing_facilities
existing facilities

existing_users
existing users

Value

tibble of users, with distances between each user and facility

Examples

# NOT RUN {
mc_facilities_selected <-extract_facility_selected(
  solution_vector = x$lp_solution$solution,
  A_mat = x$A,
  proposed_facilities = x$proposed_facility)

augmented_users <- augment_user(
    facilities_selected = mc_facilities_selected,
    existing_facilities = mc_cv_fit_n20_test_1$existing_facility,
    existing_users = mc_cv_fit_n20_test_1$existing_user
    )
# }