(Vectorized) Solve the Maximal Covering Location Problem

max_coverage_vec does what max_coverage does, but it allows for input of vectors in n_added. This function is currently still under development, as I still need to work out how to understand it.

max_coverage_vec(existing_facility = NULL, proposed_facility, user,
  distance_cutoff, n_added, solver = "lpSolve")

Arguments

existing_facility

data.frame containing the facilities that are already in existing, with columns names lat, and long.

proposed_facility

data.frame containing the facilities that are being proposed, with column names lat, and long.

user

data.frame containing the users of the facilities, along with column names lat, and long.

distance_cutoff

numeric indicating the distance cutoff (in metres) you are interested in. If a number is less than distance_cutoff, it will be 1, if it is greater than it, it will be 0.

n_added

the maximum number of facilities to add.

solver

character default is lpSolve, but glpk and Gurobi can also be used.

Value

dataframe of results