maxcovr: A Set of Tools For Solving The Maximal Covering Location Problem

Solving the "maximal covering location problem" as described by Church can be difficult for users not familiar with linear programming. maxcovr provides functions to make it easy to solve this problem, and tools to calculate facility coverage.

See also

Author

Maintainer: Nicholas Tierney nicholas.tierney@gmail.com (ORCID)

Authors:

Mark Padgham mark.padgham@email.com