Extract the number of facilities installed

Using the model-modified dataframe of proposed_facility, count the number of events installed.

Usage

n_installed(proposed_facility)

Arguments

proposed_facility

dataframe from the mc_model, of facilities proposed with the additional information about whether the facility was installed or not - is_installed

Value

datafrmae

Examples