Calculate the distance between two locations

This function uses the haversine formula to calculate the great circle distance between two locations, identified by their latitudes and longitudes. It is borrowed from rnoaa (https://github.com/ropenscilabs/rnoaa/blob/master/R/meteo_distance.R) and included here as rnoaa is a large package that is rather unrelated to maxcovr. I have renamed it from meteo_spherical_distance to spherical_distance

spherical_distance(lat1, long1, lat2, long2)

Arguments

lat1

Latitude of the first location.

long1

Longitude of the first location.

lat2

Latitude of the second location.

long2

Longitude of the second location.

Value

A numeric value giving the distance in meters between the pair of locations.

Note

This function assumes an earth radius of 6,371 km.

Examples

spherical_distance(lat1 = -27.4667,
                   long1 = 153.0217,
                   lat2 = -27.4710,
                   long2 = 153.0234)

#> [1] 506.7013