Solve the Maximal Covering Location Problem

max_coverage solves the binary optimisation problem known as the "maximal covering location problem" as described by Church (http://www.geog.ucsb.edu/~forest/G294download/MAX_COVER_RLC_CSR.pdf). This package was implemented to make it easier to solve this problem in the context of the research initially presented by Chan et al (http://circ.ahajournals.org/content/127/17/1801.short) to identify ideal locations to place AEDs.

max_coverage(existing_facility = NULL, proposed_facility, user,
  distance_cutoff, n_added, solver = "lpSolve", return_early = FALSE)

Arguments

existing_facility

data.frame containing the facilities that are already in existing, with columns names lat, and long.

proposed_facility

data.frame containing the facilities that are being proposed, with column names lat, and long.

user

data.frame containing the users of the facilities, along with column names lat, and long.

distance_cutoff

numeric indicating the distance cutoff (in metres) you are interested in. If a number is less than distance_cutoff, it will be 1, if it is greater than it, it will be 0.

n_added

the maximum number of facilities to add.

solver

character default is lpSolve, but glpk and Gurobi can also be used.

return_early

logical - should I return the object early?

Value

dataframe of results

Examples

library(dplyr)

# already existing locations
york_selected <- york %>% filter(grade == "I")

# proposed locations
york_unselected <- york %>% filter(grade != "I")

mc_result <- max_coverage(existing_facility = york_selected,
                         proposed_facility = york_unselected,
                         user = york_crime,
                         distance_cutoff = 100,
                         n_added = 20)

mc_result

#> 
#> ------------------------------------------- 
#> Model Fit: maxcovr fixed location model 
#> ------------------------------------------- 
#> model_used:        max_coverage 
#> existing_facility: york_selected 
#> proposed_facility: york_unselected 
#> user:              york_crime 
#> distance_cutoff:   100 
#> n_added:           20 
#> solver:            lpSolve 
#> -------------------------------------------

summary(mc_result)

#> 
#> ------------------------------------------- 
#> Model Fit: maxcovr fixed location model 
#> ------------------------------------------- 
#> Distance Cutoff: 100m 
#> Facilities: 
#>     Added:       20 
#> Coverage (Previous): 
#>     # Users:     540    (339) 
#>     Proportion:  0.2977 (0.1869) 
#> Distance (m) to Facility (Previous): 
#>     Avg:         986 (1400) 
#>     SD:          1188 (1597) 
#> -------------------------------------------

 # get the facilities chosen
 mc_result$facility_selected

#> [[1]]
#> # A tibble: 20 × 7
#>         long      lat object_id desig_id pref_ref
#>        <dbl>    <dbl>     <int>    <chr>    <int>
#> 1  -1.089571 53.96492      5978  DYO1383   462917
#> 2  -1.082308 53.96665      5909  DYO1297   463072
#> 3  -1.079160 53.95938      5872  DYO1244   463186
#> 4  -1.084051 53.95943      5847  DYO1216   463242
#> 5  -1.124819 53.95513      5759  DYO1108   463434
#> 6  -1.083679 53.96460      5748  DYO1096   463469
#> 7  -1.084106 53.96387      5745  DYO1093   463465
#> 8  -1.075168 53.95522      5739  DYO1086   463457
#> 9  -1.094967 53.95436      5638   DYO957   463691
#> 10 -1.092769 53.94116      5606   DYO920   463771
#> 11 -1.064879 53.95465      5592   DYO903   463788
#> 12 -1.079478 53.96778      5529   DYO829   463938
#> 13 -1.075726 53.96442      5454   DYO705   464133
#> 14 -1.115550 53.96251      5349   DYO572   464451
#> 15 -1.084055 53.95650      3991   DYO544   464508
#> 16 -1.092161 53.95868      3863   DYO406   464772
#> 17 -1.098576 53.95097      3250   DYO310   464928
#> 18 -1.001037 53.99492      3215  DYO1581   491367
#> 19 -1.079710 53.96573      3213  DYO1580   490659
#> 20 -1.083844 53.95258      4803  DYO1734       NA
#> # ... with 2 more variables: name <chr>, grade <chr>
#> 

 # get the users affected
 mc_result$user_affected

#> [[1]]
#> # A tibble: 201 × 16
#>    user_id user_chosen facility_id   distance              category
#>      <dbl>       <dbl>       <dbl>      <dbl>                 <chr>
#> 1        1           1          66   165.8575 anti-social-behaviour
#> 2        5           1          25   535.7871 anti-social-behaviour
#> 3       13           1           6  2334.3248 anti-social-behaviour
#> 4       14           1           2 12686.9300 anti-social-behaviour
#> 5       15           1          24   620.2091 anti-social-behaviour
#> 6       23           1          NA         NA                  <NA>
#> 7       27           1          NA         NA                  <NA>
#> 8       29           1           2   374.9738 anti-social-behaviour
#> 9       33           1          NA         NA                  <NA>
#> 10      48           1           6  7274.0229 anti-social-behaviour
#> # ... with 191 more rows, and 11 more variables: persistent_id <chr>,
#> #   date <chr>, lat <dbl>, long <dbl>, street_id <chr>, street_name <chr>,
#> #   context <chr>, id <chr>, location_type <chr>, location_subtype <chr>,
#> #   outcome_status <chr>
#> 

 # get the summaries
 mc_result$summary

#> [[1]]
#> # A tibble: 2 × 8
#>   n_added distance_within n_cov   pct_cov n_not_cov pct_not_cov  dist_avg
#>     <dbl>           <dbl> <int>     <dbl>     <int>       <dbl>     <dbl>
#> 1       0             100   339 0.1868798      1475   0.8131202 1400.1918
#> 2      20             100   540 0.2976847      1274   0.7023153  985.8907
#> # ... with 1 more variables: dist_sd <dbl>
#>