Find the nearest lat/long to another lat/long

This function finds the nearest lat/long pairs to another lat/long pair. So in the york building and york crime context, writing nearest(york_crime,york) reads as "find the nearest crime in york to each building in york, and returns a dataframe with every building in york, the nearest york_crime to each building, and the distance in metres between the two. Likewise, you could write nearest(york, york_crime), and this would return the nearest building to every crime. nearest assumes that the names of the latitude and longitude are "lat" and "long", but you can provide these names.

nearest(nearest_df, to_df, nearest_lat = "lat", nearest_long = "long",
  to_lat = "lat", to_long = "long")

Arguments

nearest_df

a dataframe containing latitude and longitude

to_df

a dataframe containing latitude and longitude

nearest_lat

name of latitude in nearest_df

nearest_long

name of longitude in nearest_df

to_lat

name of latitude in to_df

to_long

name of longitude in to_df

Value

dataframe of "to_df" along with the nearest "nearest_df" to each row, along with the distance between the two, and the nearest_id, the row position of the nearest_df closest to that row.

Examples

library(maxcovr)

nearest(nearest_df = york_crime,
        to_df = york)

#> # A tibble: 2,944 × 22
#>    to_id nearest_id   distance   long_to   lat_to object_id desig_id pref_ref
#>    <dbl>      <dbl>      <dbl>     <dbl>    <dbl>     <int>    <chr>    <int>
#> 1      1         33  35.953830 -1.085647 53.95965      6144  DYO1195   463280
#> 2      2        183   4.302693 -1.085600 53.96802      6143  DYO1746       NA
#> 3      3        183  35.791294 -1.085433 53.96776      6142  DYO1373   462942
#> 4      4        183  89.147429 -1.084333 53.96777      6141  DYO1745       NA
#> 5      5        942 236.154961 -1.095325 53.92372      6140    DYO14   325952
#> 6      6        942 236.154961 -1.095325 53.92372      6139    DYO14   325952
#> 7      7        245  60.566516 -1.079503 53.96029      6138  DYO1226   463228
#> 8      8        264  60.889289 -1.079771 53.96004      6137  DYO1226   463228
#> 9      9       1106  91.797647 -1.083875 53.95547      6136   DYO421   464726
#> 10    10        184  23.922817 -1.083813 53.96224      6135   DYO978   463684
#> # ... with 2,934 more rows, and 14 more variables: name <chr>, grade <chr>,
#> #   category <chr>, persistent_id <chr>, date <chr>, lat_nearest <dbl>,
#> #   long_nearest <dbl>, street_id <chr>, street_name <chr>, context <chr>,
#> #   id <chr>, location_type <chr>, location_subtype <chr>, outcome_status <chr>

# you can use the pipe as well

## Not run: ------------------------------------
# 
# library(magrittr)
# york_crime %>% nearest(york)
# 
## ---------------------------------------------